move output functions into dataflow::Result
